requirements.txt
Flask==3.1.3
pyswisseph==2.10.3.2
pytz==2026.1.post1
gunicorn==25.1.0
timezonefinder>=8.0.0
astral==3.2
requests==2.32.5
